шаблонизируем проект
