+++ data.php
@@ -1,0 +1,47 @@
+<?php
+$is_auth = rand(0, 1);
+
+$user_name = 'Татьяна'; // укажите здесь ваше имя
+
+$title = 'Главная';
+
+$categories = ["Доски и лыжи", "Крепления", "Ботинки", "Одежда", "Инструменты", "Разное"];
+
+$lots = [
+    [
+        'name' => '2014 Rossignol District Snowboard',
+        'category' => 'Доски и лыжи',
+        'cost' => 10999,
+        'url' => 'img/lot-1.jpg',
+    ],
+    [
+        'name' => 'DC Ply Mens 2016/2017 Snowboard',
+        'category' => 'Доски и лыжи',
+        'cost' => 159999,
+        'url' => 'img/lot-2.jpg',
+    ],
+    [
+        'name' => 'Крепления Union Contact Pro 2015 года размер L/XL',
+        'category' => 'Крепления',
+        'cost' => 8000,
+        'url' => 'img/lot-3.jpg',
+    ],
+    [
+        'name' => 'Ботинки для сноуборда DC Mutiny Charocal',
+        'category' => 'Ботинки',
+        'cost' => 10999,
+        'url' => 'img/lot-4.jpg',
+    ],
+    [
+        'name' => 'Куртка для сноуборда DC Mutiny Charocal',
+        'category' => 'Одежда',
+        'cost' => 7500,
+        'url' => 'img/lot-5.jpg',
+    ],
+    [
+        'name' => 'Маска Oakley Canopy',
+        'category' => 'Разное',
+        'cost' => 5400,
+        'url' => 'img/lot-6.jpg',
+    ],
+];
